for default tests:
**before:**
- Latency: 0.48080000281333923 ms
- TFlops: 53.59776128371738 TFlops

**after:**
- Latency: 0.30057600140571594 ms
- TFlops: 85.73473482740243 TFlops

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- Refactor
  - Kernel/grid sizing now uses a precomputed total M-block count for more consistent launch dimensions.
  - Batching and padding logic streamlined while preserving overall behavior.
- Bug Fixes
  - Fixed padding calculation to prevent unintended extra padding in batched GEMM.
- Notes
  - No API changes; behavior remains compatible.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->